Improve ActivityLifecycle and FragmentLifecycle
