+++ arms/src/main/java/com/jess/arms/integration/FragmentLifecycle.java
@@ -1,0 +1,135 @@
+package com.jess.arms.integration;
+
+import android.content.Context;
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.view.View;
+
+import com.jess.arms.base.delegate.FragmentDelegate;
+import com.jess.arms.base.delegate.FragmentDelegateImpl;
+import com.jess.arms.base.delegate.IFragment;
+
+import timber.log.Timber;
+
+/**
+ * ================================================
+ * Created by JessYan on 04/09/2017 16:04
+ * Contact with jess.yan.effort@gmail.com
+ * Follow me on https://github.com/JessYanCoding
+ * ================================================
+ */
+
+public class FragmentLifecycle extends FragmentManager.FragmentLifecycleCallbacks {
+
+    @Override
+    public void onFragmentAttached(FragmentManager fm, Fragment f, Context context) {
+        Timber.w(f.toString() + " - onFragmentAttached");
+        if (f instanceof IFragment && f.getArguments() != null) {
+            FragmentDelegate fragmentDelegate = fetchFragmentDelegate(f);
+            if (fragmentDelegate == null || !fragmentDelegate.isAdded()) {
+                fragmentDelegate = new FragmentDelegateImpl(fm, f);
+                f.getArguments().putParcelable(FragmentDelegate.FRAGMENT_DELEGATE, fragmentDelegate);
+            }
+            fragmentDelegate.onAttach(context);
+        }
+    }
+
+    @Override
+    public void onFragmentCreated(FragmentManager fm, Fragment f, Bundle savedInstanceState) {
+        Timber.w(f.toString() + " - onFragmentCreated");
+        FragmentDelegate fragmentDelegate = fetchFragmentDelegate(f);
+        if (fragmentDelegate != null) {
+            fragmentDelegate.onCreate(savedInstanceState);
+        }
+    }
+
+    @Override
+    public void onFragmentViewCreated(FragmentManager fm, Fragment f, View v, Bundle savedInstanceState) {
+        Timber.w(f.toString() + " - onFragmentViewCreated");
+        FragmentDelegate fragmentDelegate = fetchFragmentDelegate(f);
+        if (fragmentDelegate != null) {
+            fragmentDelegate.onCreateView(v, savedInstanceState);
+        }
+    }
+
+    @Override
+    public void onFragmentActivityCreated(FragmentManager fm, Fragment f, Bundle savedInstanceState) {
+        Timber.w(f.toString() + " - onFragmentActivityCreated");
+        FragmentDelegate fragmentDelegate = fetchFragmentDelegate(f);
+        if (fragmentDelegate != null) {
+            fragmentDelegate.onActivityCreate(savedInstanceState);
+        }
+    }
+
+    @Override
+    public void onFragmentStarted(FragmentManager fm, Fragment f) {
+        Timber.w(f.toString() + " - onFragmentStarted");
+        FragmentDelegate fragmentDelegate = fetchFragmentDelegate(f);
+        if (fragmentDelegate != null) {
+            fragmentDelegate.onStart();
+        }
+    }
+
+    @Override
+    public void onFragmentResumed(FragmentManager fm, Fragment f) {
+        Timber.w(f.toString() + " - onFragmentResumed");
+        FragmentDelegate fragmentDelegate = fetchFragmentDelegate(f);
+        if (fragmentDelegate != null) {
+            fragmentDelegate.onResume();
+        }
+    }
+
+    @Override
+    public void onFragmentPaused(FragmentManager fm, Fragment f) {
+        Timber.w(f.toString() + " - onFragmentPaused");
+        FragmentDelegate fragmentDelegate = fetchFragmentDelegate(f);
+        if (fragmentDelegate != null) {
+            fragmentDelegate.onPause();
+        }
+    }
+
+    @Override
+    public void onFragmentStopped(FragmentManager fm, Fragment f) {
+        Timber.w(f.toString() + " - onFragmentStopped");
+        FragmentDelegate fragmentDelegate = fetchFragmentDelegate(f);
+        if (fragmentDelegate != null) {
+            fragmentDelegate.onStop();
+        }
+    }
+
+    @Override
+    public void onFragmentViewDestroyed(FragmentManager fm, Fragment f) {
+        Timber.w(f.toString() + " - onFragmentViewDestroyed");
+        FragmentDelegate fragmentDelegate = fetchFragmentDelegate(f);
+        if (fragmentDelegate != null) {
+            fragmentDelegate.onDestroyView();
+        }
+    }
+
+    @Override
+    public void onFragmentDestroyed(FragmentManager fm, Fragment f) {
+        Timber.w(f.toString() + " - onFragmentDestroyed");
+        FragmentDelegate fragmentDelegate = fetchFragmentDelegate(f);
+        if (fragmentDelegate != null) {
+            fragmentDelegate.onDestroy();
+        }
+    }
+
+    @Override
+    public void onFragmentDetached(FragmentManager fm, Fragment f) {
+        Timber.w(f.toString() + " - onFragmentDetached");
+        FragmentDelegate fragmentDelegate = fetchFragmentDelegate(f);
+        if (fragmentDelegate != null) {
+            fragmentDelegate.onDetach();
+            f.getArguments().clear();
+        }
+    }
+
+    private FragmentDelegate fetchFragmentDelegate(Fragment fragment) {
+        if (fragment instanceof IFragment) {
+            return fragment.getArguments() == null ? null : fragment.getArguments().getParcelable(FragmentDelegate.FRAGMENT_DELEGATE);
+        }
+        return null;
+    }
+}
